+++ examples/data-sources/automox_organizations/data-source.tf
@@ -1,0 +1,9 @@
+data "automox_organizations" "all" {}
+
+# How close the organization is to its licensed device limit.
+output "device_headroom" {
+  value = {
+    for o in data.automox_organizations.all.organizations : o.name
+    => o.device_limit - o.device_count
+  }
+}